requirements.txt
qlik-sdk==0.17.0
load_dotenv
requests